fix bug related to task ui state reference
