+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/taskDetalis/TaskViewDetails.kt
@@ -27,7 +27,7 @@ import com.sanaa.tudee_assistant.presentation.design_system.theme.Theme
 import com.sanaa.tudee_assistant.presentation.design_system.theme.TudeeTheme
 import com.sanaa.tudee_assistant.presentation.model.TaskUiPriority
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
-import com.sanaa.tudee_assistant.presentation.screen.taskScreen.TaskUiModel
+import com.sanaa.tudee_assistant.presentation.state.TaskUiModel
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
