+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/taskScreen/TaskStatusTabs.kt
@@ -12,6 +12,7 @@ import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.design_system.component.TabItem
 import com.sanaa.tudee_assistant.presentation.design_system.component.TudeeTabs
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
+import com.sanaa.tudee_assistant.presentation.state.TaskUiModel
 
 @Composable
 fun TaskStatusTabs(
