+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/taskScreen/TasksViewModel.kt
@@ -6,6 +6,9 @@ import com.sanaa.tudee_assistant.domain.model.Category
 import com.sanaa.tudee_assistant.domain.service.CategoryService
 import com.sanaa.tudee_assistant.domain.service.TaskService
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
+import com.sanaa.tudee_assistant.presentation.screen.taskScreen.mapper.toTask
+import com.sanaa.tudee_assistant.presentation.screen.taskScreen.mapper.toUiModel
+import com.sanaa.tudee_assistant.presentation.state.TaskUiModel
 import com.sanaa.tudee_assistant.presentation.utils.BaseViewModel
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
